#5 Small refactor to make screens more testable
